change Box -> Polymorphic
