+++ packages/itwinui-react/src/core/utils/components/index.ts
@@ -16,4 +16,4 @@ export * from './Flex.js';
 export * from './Divider.js';
 export * from './LinkAction.js';
 export * from './AutoclearingHiddenLiveRegion.js';
-export * from './Box.js';
+export * from './Polymorphic.js';
